chore: renamed apiJsonPaths to sources for parser configuration
