+++ libs/ui5-angular/wrapper.conf.ts
@@ -52,7 +52,7 @@ const packageNames = ['fiori', 'main'];
 
 const ui5WrapperConfig: Ui5NxWrapperConfig<ComponentData> = {
   getComponents: () => apiJsonParser({
-    apiJsonPaths: [
+    sources: [
       '@ui5/webcomponents-base/dist/api.json',
       '@ui5/webcomponents-fiori/dist/api.json',
       '@ui5/webcomponents/dist/api.json'
