+++ libs/ui5-schema-parser/src/lib/api-json-parser/get-components.ts
@@ -1,6 +1,6 @@
 import {SymbolObject} from "./symbol-types";
 import {camelCase, kebabCase} from 'lodash';
-import {ComponentData, ParameterType} from "@ui5/webcomponents-wrapper";
+import {ComponentData} from "@ui5/webcomponents-wrapper";
 
 const objectTypeMapper = (elementTagName: string, identifier: string) => {
   if (elementTagName === 'ui5-button') {
