+++ libs/ui5-schema-parser/src/lib/custom-element-schema-parser/index.ts
@@ -1,0 +1,11 @@
+import { ComponentData } from "@ui5/webcomponents-wrapper";
+import { ParserConfiguration } from "../parser-configuration";
+
+/**
+ * Parses the custom-elements.json files and returns the list of the component data
+ */
+export async function customElementSchemaParser(options: ParserConfiguration): Promise<ComponentData[]> {
+  const components: ComponentData[] = [];
+
+  return components;
+}
